✨ Buildings addressable by url
